==> notebooks/2_data_preprocessing.ipynb <==
import pandas as pd
import sys
from icecream import ic
from sklearn.preprocessing import StandardScaler
sys.path.append("./../projectname")
from projectname.config import raw_data_dir
from projectname.config import clean_data_dir

# Load the dataset
bookings_path = raw_data_dir / 'hotel_bookings.csv'
bookings_df = pd.read_csv(bookings_path)
bookings_df

# Handle Missing Values

missing_values = bookings_df.isnull().sum()
missing_values = missing_values[missing_values > 0]
ic(missing_values)

# Step 1: Handle Missing Values
bookings_df['children'] = bookings_df['children'].fillna(bookings_df['children'].median())
bookings_df['country'] = bookings_df['country'].fillna(bookings_df['country'].mode()[0])
bookings_df['agent'] = bookings_df['agent'].fillna(-1)  # Using -1 to indicate missing agent
bookings_df['company'] = bookings_df['company'].fillna(-1)  # Using -1 to indicate missing company
bookings_df


# Step 2: Convert Categorical Variables
# Drop 'reservation_status_date' before encoding
bookings_df.drop(columns=['reservation_status_date'], inplace=True)

categorical_cols = bookings_df.select_dtypes(include=['object']).columns
categorical_cols

bookings_df.to_csv(clean_data_dir / "bookings_clean.csv")

# Perform one-hot encoding on categorical columns
bookings_df = pd.get_dummies(bookings_df, columns=categorical_cols, drop_first=True)
bookings_df

# Step 4: Normalize/Scale Features
scaler = StandardScaler()
scaled_columns = ['lead_time', 'arrival_date_week_number', 'arrival_date_day_of_month',
                  'stays_in_weekend_nights', 'stays_in_week_nights', 'adults', 'children',
                  'babies', 'adr', 'days_in_waiting_list', 'required_car_parking_spaces',
                  'total_of_special_requests']

bookings_df[scaled_columns] = scaler.fit_transform(bookings_df[scaled_columns])
bookings_df

# Step 5: Remove Duplicates
bookings_df = bookings_df.drop_duplicates()
bookings_df

bookings_clean_path = clean_data_dir / 'bookings.csv'
bookings_df.to_csv(bookings_clean_path)
