==> notebooks/4_model_testing.ipynb <==
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
import joblib
import pandas as pd
from sklearn.pipeline import Pipeline

# Load the saved model
pipeline_filename = "../results/random_forest_model_pipeline.pkl"
loaded_pipeline = joblib.load(pipeline_filename)

test_set_filename = '../data/clean/X_test.csv'
loaded_X_test = pd.read_csv(test_set_filename)



# Use the loaded pipeline to make predictions
sample_data = loaded_X_test.iloc[0:1]  # Take the first sample from the test set
predicted_class = loaded_pipeline.predict(sample_data)

print(f"Predicted class: {predicted_class[0]}")


# Generate predictions for the entire test set using the Random Forest model
predicted_classes = loaded_pipeline.predict(loaded_X_test)

# Create a dataframe with the X_test and add the predicted class as a new columns
X_test_with_predictions = loaded_X_test.copy()
X_test_with_predictions['predicted_class'] = predicted_classes

X_test_with_predictions
